RTS fix for x86_64 architecture.
John and I discovered that the generated C code does not run on 64bit architectures.
Therefore we made the following changes:
(*) Changed type of a GrWord from unsigned int to intptr_t. This latter type is just an
    integer with the additional guarantee that you can store a pointer in it (and get
    it back untouched) [size is 4 for 32bit architectures and 8 for 64bit architectures].
    This type is defined in <inttypes.h> and should be portable.
(*) (difficult to find bug) added a declaration for heapalloc() to rts.h. Originally, this
    declaration wasn't there, however, the C compiler assumes that the return value of
    an undeclared function is an int.
(*) changed the configure.ac to detect the x86_64 architecture to properly generate
    ASM for tailcalls.
(*) (cosmetic change) added an import of <stdlib.h> to get rid of undeclared malloc
    warnings.