Full prelude, and more of John's benchmarks survice HPT analysis.
Cause of much trouble in HPT analysis was that the input Grin code is too liberal on the use of variables for containing nodes or pointers. In the analysis, we need parameters always to be a pointer, binds always to return a node, etc.
Transformation "AddFetch" tries to accomplish this invariant. I extended it to cover some cases hitherto neglected.

I also changed the prelude in removing primShowInt, because primitives
with String result are taboo in my world. I replaced it by a version written in Haskell, to be replaced some day be a version returning a PackedString.

Consequently, we can now HPT-analyse the full Prelude.hs, and test t1.hs.
The previous version, that is, because commit v.934 introduces new problems.

After all program transformations are done, the program hangs;
so there is still a problem in generating Silly code.

But we can compile and run more of John's benchmarks:
* exp3_8 already worked (answers 6561 in 1 sec.)
* tak already worked (answers 9 in 78 sec.)
* primes now also compiles.
  - on input 5500 it runs out of memory
  - on input 500 it segfaults
* queens was slightly rewritten to avoid literal patterns.
  it now runs, and on input 12 answers 14200 in 63 sec.
* wheel-sieve2 was also rewritten without literal patterns, but still has a HPT problem.

Other nuisances:
Major: "ehc -cbexe t1.exe" generates code, but crashes when run on my PC
Minor: when "ehc -cexe" is interupted (which is currently always the case),
it cannot be run again, bacuse "oneBigProgram" panics. If you run "ehc -cbexe" in between, the state is restored.