Fixed a terrible flaw in HPT analysis, which caused problems when using #A on expressions that return #P functions. Introduced a Grin "mode-checker", that checks the "Grin mode invariant": whether functions are called with pointer arguments only. Prepared many Grin transformations for the forthcoming change in syntax of Grin lambda-patterns.
Many functions from nofib which formerly had problems now compile and run: digits-of-e2, primes, wheel-sieve1.

Open problems:
Wheel-sieve1 does not produce output on input >10000. This seems to be an overflow problem rather than HPT-related.
Wheel-sieve2 still does not compile. This is due to a violation of the "Grin mode invariant" (detected by the new checker),
which occurs in the generated code from Core2Grin for the "span" prelude function.