Fixed an error in Silly when copying stack frames on tail calls.
The SmartAssignment strategy is capable of copying overlapping stack frames. It can do simple copies like
  SP[1] = SP[5]
  SP[2] = SP[6]
it orders assignments when necessary, as in
  SP[3] = SP[2]
  SP[2] = SP[1]
and introduces an auxiliary variable if assignments are mutually dependent:
  AUX = SP[1]
  SP[1] = SP[2]
  SP[2] = AUX
Also, it eliminates trivial assignments like
  SP[2] = SP[2]
The error was that it eliminated assignments of which the destination was *dependent* on the source, istead of when it was *equal* to the source. Therefore, it also eliminated
  SP[2] = SP[2][1]

This error happened for the first time now, because the specialization of the app function for each occurence of #A tags in grin introduced many more opportunities for this optimization.
It is a good thing that these optimizations are possible, and they are done right now.