Fixed a problem with SplitFetch and CopyPropagation which arose as a result of the recent introduction of app-function specialization.
The app-function specialization results in case-distinctions with fewer arms (which is a good thing!). When only one arm remains, the case-expression is eliminated entirely and the eval which precedes it is replaced by a simple fetchNode. For example:
    eval x; \y ->
    case y of
      (C: a b) -> ...
is replaced by
    fetchnode x; \(C: a b) -> ...
Now there can be additional Unit's between Eval and Case, for example:
    eval x; \y ->
    unit y; \z ->
    case z of
      (C: a b) -> ...
in which case we get:
    fetchnode x; \(p q r) ->
    unit (p q r); \(C: a b) -> ...

Originally SplitFetch only fetches all fields if the tag is known. Otherwise, it assumes that a Case-expression follows, in which all but the first fields can be fetched.
In this situation described, however, SplitFetch was not smart enough to see that p can be only C: and thus the fetching of q and r was postponed to the subsequent Case-statement that is not there anymore.

As a fix, I adapted CopyPropagation to replace
    fetchnode x; \(p q r) ->
    unit (p q r); \(C: a b) -> ...
not by
    fetchnode x; \(p q r) -> ...[q/a, r/b]
but by
    fetchnode x; \(C: q r) -> ...[q/a, r/b]
Now the SplitFetch works fine.

As a result, the benchmark nofib/spectral/scc now compiles.
At runtime, however, it encounters a black hole.
(in the bexe backend too, it wrongly halts with "0" result).