A new transformation merges adjacent Grin case-expressions in some situations. Many such opportunities arise from the recently introduced "Late Inlining". Currently it is only performed with --priv=1, because the LLVM backend is not yet ready for it.
Detailed explanation:
Many case expression in Grin originate from inlining Eval.
They have the following form:
    CASE x
    {   [N] #F/f y -> CALL f y; \r ->
                      UPDATEUNIT r x
        [I] #C/c   -> UNIT x
    }; \z ->
That is: if a function was postponed, it is called now, and the result updates the redex;
if it was already evaluated, nothing happens.
Often, such a case expression is followed by another case:
    CASE z
    {   [N] #C/c   -> do something useful
    }
It is a waste do do the case-distinction twice.
As a preparation to the merge, we annotate the alternatives as either "[N]ormal" or "[I]dent".

Two adjacent case-expressions can be merged if the pattern after the first (z in the example)
is the scrutinee of the second.
In the example we get:
    CASE x
    {   [R] #F/f y -> CALL f y; \r ->
                      UPDATEUNIT r x
        [N] #C/c   -> do something useful
    }
That is: the [N]ormal alternatives from the first case are re-annotated as [R]e-entrant;
the [I]dent alternatives from the first case are thrown away.
All alternatives from the second case are kept.
Alternatives annotated [R]e-entrant have special semantics: they generate special code, 
in order to scrutinize the result again.

Subtlety: any z that may occur free in the body must be replaced by x.