Solved two problems with Grin transformations "Inline" and "MergeCase".
"Inline" was developed for the bytecode backend. It can however also be used
in the compiled backends, provided that it doe not remove the header of the inlined function. The is a boolean parameter now that controls this.

"MergeCase" assumed that the "normal" alternatives always have a non-final tag.
John found a situation where this is not the case (the situation only occurs if
both the Inline and the EvalStore optimizations are done). The condition is explicitly
checked now.