- work on deriving for Read - refactoring fragments of Ty/FitsIn, for - bugfix for incorrect instantiation for predicative binding in function application, for example in:
ids :: [forall a . a -> a]

h4 = tail ids
h5 = id   ids

would previously be inferred

h4 :: forall a . [a -> a]
h5 :: [forall a . a -> a]

'forall a . [a -> a]' enforced by the type signature of the argument of tail, but not by the argument type of id.
This is inconsistent, or at least difficult to explain, so the default is now deep instantiation for all covariant arguments of a type constructor.
This leads to:

h4, h5 :: forall a . [a -> a]

unless impredicative application is explicitly specified:

h4 = tail ~ids
h5 = id   ~ids

which leads to

h4, h5 :: [forall a . a -> a]