Cleanup in the way the HPT table is handled.
The HPT table is created by PointToAnalysis, and later updated by all transformations that introduce new variables (InlineEA, LateInline, LowerGrin and SplitFetch). These transformations now do so in a uniform way (collect modifications and insert them after the tree walk). Some of them used to thread the HPT-table through the treewalk and modify it on the fly.

As a consequence, we can store the table now as a single Array (it used to consist of an Array as a base part, and a Map with the modifications).

As a consequence, we don't need to store the first unused unique number explicitly any more: it is removed from the compiler state, and and passed to the transformations anymore.