More libraries: Numeric + others required for this one
Fixed:
- incorrect occur check failure (infinite type) in case expression alternatives, a corner case combi with implicit parameters, do notation, class Num.
- forgotten: context reduction did use instances also for predicates over generalized type variables, thereby closing the world of instances too early.

A bit of preparation to separate the notion of substitution and substitutable, to allow for flexibility in substutions,
most likely required for default instances (because context reduction then narrows a type in a class predicate to its default),
surely required for later work (e.g. functional dependencies, where type improvements will have to be propagated by means of substitutions).