Fix of variant 8 faulty C generation for interpreter.
Library for arrays, almost done.

Related to arrays: primitives did require annotation for FFI calls, to
know whether an ffi returns an evaluated value. By default this is true,
but for the primitive reading a value from an array this is not, because
arrays may store unevaluated thunks. For now it is a bit of ad-hoc
solution, and the annotation really only is required for Grin code
generation (because an explicit eval is required after the call) and by
the eval elimination transformation (because it can only assume
evaluation depending on the annotation). It has to be checked with full
prog analysis and eval inlining.

Fix of combination of a strict binding (via letstrict) and additional
bindings for proven class predicates. The binding order here matters
because there is an explicit evaluation order, but bindings were mixed
up. Now the two kinds of bindings (strict and non strict) are separated
and bound in the correct order. The problem did not arise until now
because the bound expression usually consisted of a single non
overloaded identifier. In the Data.Array library however, much work is
done in the ST monad and requires explicit evaluation.