For posterity, tagged revision 1522, the version of the EHC that has a compiling CLR backend, originally released March 19th, 2009. This is a pre UHC 1.0.0 version of the code.
This backend was developed by Jeroen Leeuwestein and Tom Lokhorst in early 2009 as part of the Advanced Compiler Construction course.
To assemble and run CLR code, Mono or .NET needs to be installed. Specific details of the CLR backend can be found in chapter 8 of the technical documentation.

Install prerequisites:
    $ cd extlibs/cil/
    $ tar xzf language-cil-0.1.0.tar.gz
    $ cd language-cil-0.1.0
    $ cabal install

Compile EHC as follows:
    $ ./configure --enable-clr
    $ make 8/ehc

Run test as follows:
   $ cd test/clr/1-add/
   $ ../../../install/8/bin/ehc -tclr Test.hs
   $ ilasm2 Test.il
   $ mono Test.exe
   Int 42