Fixed various potential and actual GC bugs. The problem arises at various locations in the RTS and (probably) also in generated code when an indirection node is accessed whilst assuming it is an evaluated node. The RTS has been scrutinized, assertion code inserted to check this. The generated code and transformations yet have to be checked for this.
The problem does not arise in the regression tests, only with nofib test programs, and only in longer runs. Downsizing the admin and memory blocksize used for GC seems to avoid the problem. It may well be necessary to first instrument the RTS further because the generated dumps are too large to produce and look at (at least > .5 GB, taking minutes & GB's to be produced).