Fixed various problems related to elimination of GRIN name aliases (AliasElim) and removal of redundant evaluations (EvalElim): - EvalElim did move evaluations in front of others, this broke `seq` usage. - EvalElim incorrectly assumed evaluation of applications in some cases. - AliasElim did replace aliases at the tail of a function incorrectly involving an eval at the end - AliasElim did also deal with the EvalElim's task.
From the benchmark tests, paraffins now runs without stack overflow, as it should, but for large runs it runs out of heap. Reason still unknown, but a good guess is the use of arrays which are still used without finalizer (because GC finalization is not yet implemented).

As a side effect, several nofib tests run faster.