The C backend generated invalid C code: CUHC_IOBase_ReadMode was used, but not defined. A comment in the tagConstants function in GrinCode/ToSilly suggested the fix:
-- Make sure these are the same as in src/rts/C/prim-const.ch

The names in that file were of the form "C$name" (i.e. CReadMode). I copied the
definitions, changing the naming scheme to "C_UHC_IOBase_$name". As there
already was a CUHC_IOBase_AppendBinaryMode, I think this was a safe thing to do.
I left the "#define C$name"s in there in order not to break stuff.