Fix of parsing ambiguities introduced by offside parsing from uulib.
Implementation of new rule (as per Haskell2010) for parsing do
expressions, which now does not require indentation to start of a new do
expression. In the process of this and above fix an old do related
ambiguity has now been fixed as well.

WARNING: uulib>=0.9.11, which has been uploaded to hackage
         make clean, etc