Introduces frames on the heap and meta-closures. Both work!
Major Changes:
Modifications to Grin and Silly AbsSyn to incoperate meta-closures.
Introduces a new Grin transformation: IntroMeta which introduces meta-closures by converting F-tags this is done just after the NumberIdents Grin trf.
Grin trf InlineEA produces code for the new M (Meta) tags
Grin trfs SplitFetch and LowerGrin (locally) replace variables which might be used by meta-closures so they are followed by a new transformation: SubstMeta which globally replaces variables use in meta-closures.
Silly trf EmbedVars is heavily altered: 
	- put frames on the heap instead of stack space
        - translates the new variable offsets used in meta-closures to actual offsets

Bugs/Considerations:

Apperantly the LowerGrin transformation in combination with meta-closures messes up the HPT table, because a subsequent ImpossibleCase transformation will remove to 
many cases causing the program to crash. For this reason the last ImpossibleCase trf is currently disabled. Fixing this is the next thing to do....

The program crashes as soon as the Boehm garbage collector tries to collect. This is because frames are now on the heap and are not specified as a root-set. 
This will not be fixed, instead I will make the whole thing work with the new GC.