Trigger MemberSelect with the demo code. Incrementalize MemberSelect.
This does not work correctly, because I went for a quick and ad hoc approach,
which I thought I would be able to get away with temporarily:

When doing transformation X on module B, I collect the needed information
(environments etc) from imported module A, which has already been fully
transformed. Because of later transformations, this code is not in the form
which transformation X expects, so the collected information is wrong
(incomplete in this case).

The correct way would be to store the information when doing the actual
transformation X on module A, and use that when doing transformation X on B.